coincap/fnc-coincap-history-agentfriendly/requirements.txt
pandas
requests
google-cloud-bigquery
google-cloud-secret-manager
pandas-gbq